+++ bootcamp/src/main/java/Spotify/persistence/repository/ArtistRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArtistRepository extends PagingAndSortingRepository<ArtistEntity,Integer> {
+public interface ArtistRepository extends PagingAndSortingRepository<ArtistEntity,Long> {
 
 }
