+++ bootcamp/src/main/java/Spotify/persistence/repository/GenreRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GenreRepository extends PagingAndSortingRepository<GenreEntity,Integer> {
+public interface GenreRepository extends PagingAndSortingRepository<GenreEntity,Long> {
 
 
 }
